Move languageserver and fallback to completions
