+++ spyder/plugins/completion/fallback/actor.py
@@ -26,9 +26,9 @@ from pygments.lexers import get_lexer_by_name
 from diff_match_patch import diff_match_patch
 
 # Local imports
-from spyder.plugins.languageserver import CompletionItemKind
-from spyder.plugins.languageserver import LSPRequestTypes
-from spyder.plugins.fallback.utils import get_keywords, get_words
+from spyder.plugins.completion.languageserver import CompletionItemKind
+from spyder.plugins.completion.languageserver import LSPRequestTypes
+from spyder.plugins.completion.fallback.utils import get_keywords, get_words
 
 
 logger = logging.getLogger(__name__)
