+++ spyder/plugins/completion/fallback/plugin.py
@@ -15,8 +15,8 @@ import logging
 
 # Local imports
 from spyder.api.completion import SpyderCompletionPlugin
-from spyder.plugins.fallback.actor import FallbackActor
-# from spyder.plugins.languageserver import LSPRequestTypes
+from spyder.plugins.completion.fallback.actor import FallbackActor
+# from spyder.plugins.completion.languageserver import LSPRequestTypes
 
 
 logger = logging.getLogger(__name__)
