+++ spyder/plugins/completion/fallback/tests/data/example.R
