+++ spyder/plugins/completion/fallback/tests/data/example.jl
