+++ spyder/plugins/completion/fallback/tests/data/example_latin1.py
