+++ spyder/plugins/completion/fallback/utils.py
