+++ spyder/plugins/completion/languageserver/decorators.py
@@ -7,7 +7,7 @@
 """Spyder Language Server Protocol Client auxiliar decorators."""
 
 import functools
-from spyder.plugins.languageserver.transport import MessageKind
+from spyder.plugins.completion.languageserver.transport import MessageKind
 
 
 def send_request(req=None, method=None):
