+++ spyder/plugins/completion/languageserver/providers/utils.py
