+++ spyder/plugins/completion/languageserver/providers/workspace.py
@@ -8,10 +8,11 @@
 
 import logging
 
-from spyder.plugins.languageserver.providers.utils import (
+from spyder.plugins.completion.languageserver.providers.utils import (
     path_as_uri, process_uri, match_path_to_folder)
-from spyder.plugins.languageserver import LSPRequestTypes, ClientConstants
-from spyder.plugins.languageserver.decorators import (
+from spyder.plugins.completion.languageserver import (
+    LSPRequestTypes, ClientConstants)
+from spyder.plugins.completion.languageserver.decorators import (
     handles, send_request, send_response, send_notification)
 
 logger = logging.getLogger(__name__)
