+++ spyder/plugins/completion/languageserver/tests/test_client.py
@@ -11,8 +11,8 @@ import pytest
 from qtpy.QtCore import QObject, Signal, Slot
 
 from spyder.config.lsp import PYTHON_CONFIG
-from spyder.plugins.languageserver.client import LSPClient
-from spyder.plugins.languageserver import LSPRequestTypes
+from spyder.plugins.completion.languageserver.client import LSPClient
+from spyder.plugins.completion.languageserver import LSPRequestTypes
 
 
 class CompletionPlugin(QObject):
