+++ spyder/plugins/completion/languageserver/transport/common/consumer.py
