+++ spyder/plugins/completion/languageserver/transport/common/producer.py
