+++ spyder/plugins/completion/languageserver/transport/stdio/consumer.py
@@ -15,7 +15,7 @@ client via ZMQ.
 """
 
 import logging
-from spyder.plugins.languageserver.transport.common.consumer import (
+from spyder.plugins.completion.languageserver.transport.common.consumer import (
     IncomingMessageThread)
 
 logger = logging.getLogger(__name__)
