+++ spyder/plugins/completion/plugin.py
@@ -24,10 +24,10 @@ from spyder.config.lsp import PYTHON_CONFIG
 from spyder.config.main import CONF
 from spyder.api.completion import SpyderCompletionPlugin
 from spyder.utils.misc import select_port, getcwd_or_home
-from spyder.plugins.languageserver.plugin import LanguageServerPlugin
-from spyder.plugins.fallback.plugin import FallbackPlugin
-from spyder.plugins.languageserver import LSPRequestTypes
-# from spyder.plugins.languageserver.confpage import LanguageServerConfigPage
+from spyder.plugins.completion.languageserver.plugin import (
+    LanguageServerPlugin)
+from spyder.plugins.completion.fallback.plugin import FallbackPlugin
+from spyder.plugins.completion.languageserver import LSPRequestTypes
 
 
 logger = logging.getLogger(__name__)
