+++ spyder/plugins/editor/widgets/codeeditor.py
@@ -57,9 +57,9 @@ from spyder.plugins.editor.extensions import (CloseBracketsExtension,
                                               DocstringWriterExtension,
                                               QMenuOnlyForEnter,
                                               EditorExtensionsManager)
-from spyder.plugins.languageserver import (LSPRequestTypes,
-                                           TextDocumentSyncKind,
-                                           DiagnosticSeverity)
+from spyder.plugins.completion.languageserver import (LSPRequestTypes,
+                                                      TextDocumentSyncKind,
+                                                      DiagnosticSeverity)
 from spyder.plugins.editor.panels import (ClassFunctionDropdown,
                                           DebuggerPanel, EdgeLine,
                                           FoldingPanel, IndentationGuide,
