+++ spyder/plugins/help/tests/test_plugin.py
@@ -25,7 +25,7 @@ from flaky import flaky
 
 # Local imports
 from spyder.plugins.help.plugin import Help
-from spyder.plugins.fallback.utils import default_info_response
+from spyder.plugins.completion.fallback.utils import default_info_response
 
 
 # =============================================================================
